Added in the ability to add an exercise to a workout and to reorder exercises within the workout

+++ src/main/java/com/lukeroche/fit/domain/entities/WorkoutExerciseEntity.java
@@ -15,7 +15,8 @@ import java.util.List;
 @Entity
 @Table(name = "workout_exercises")
 public class WorkoutExerciseEntity extends BaseEntity{
-    private Integer orderIndex;
+
+    private Long orderIndex;
 
     //private String notes;
